requirements.txt
pillow==10.4.0
requests==2.32.3